requirements.txt
flask
opencv-python
mediapipe
numpy
requests
python-dotenv
openai
SpeechRecognition
gTTS
pygame
rich
pyaudio
